// src/api/models/PluginUIFeature.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Serializer for a plugin ui feature.
 */
export type PluginUIFeature = {
    plugin_name: string;
    feature_type: string;
    key: string;
    title?: string;
    description?: string;
    icon?: string;
    options?: Record<string, any>;
    context?: Record<string, any>;
    source?: string;
};